+++ test/module-test.js
@@ -10,7 +10,7 @@ vows.describe('module tests').addBatch({
       return css.minify.bind(css);
     },
     'should minify CSS correctly': function(minify) {
-      assert.equal(minify('a{  color: #f00;  }'), 'a{color:red}');
+      assert.equal(minify('a{  color: #f00;  }').styles, 'a{color:red}');
     }
   },
   'extended via prototype': {
@@ -21,7 +21,7 @@ vows.describe('module tests').addBatch({
       new CleanCSS().foo('a{  color: #f00;  }', this.callback);
     },
     'should minify CSS correctly': function(error, minified) {
-      assert.equal(minified, 'a{color:red}');
+      assert.equal(minified.styles, 'a{color:red}');
     },
     teardown: function() {
       delete CleanCSS.prototype.foo;
@@ -39,7 +39,7 @@ vows.describe('module tests').addBatch({
       assert.equal(errors, null);
     },
     'should yield minified data': function(errors, minified) {
-      assert.equal(minified, 'a{color:red}');
+      assert.equal(minified.styles, 'a{color:red}');
     }
   },
   'with callback passed and one error': {
@@ -111,7 +111,7 @@ vows.describe('module tests').addBatch({
       this.callback(null, minified, minifier);
     },
     'should minify correctly': function(error, minified) {
-      assert.equal(minified, 'a{display:block}');
+      assert.equal(minified.styles, 'a{display:block}');
     },
     'should raise no errors': function(error, minified, minifier) {
       assert.equal(minifier.errors.length, 0);
@@ -128,7 +128,7 @@ vows.describe('module tests').addBatch({
       this.callback(null, minified, minifier);
     },
     'should minify correctly': function(error, minified) {
-      assert.equal(minified, 'a{display:block}p{color:red}');
+      assert.equal(minified.styles, 'a{display:block}p{color:red}');
     },
     'should raise no errors': function(error, minified, minifier) {
       assert.equal(minifier.errors.length, 0);
@@ -145,7 +145,7 @@ vows.describe('module tests').addBatch({
       this.callback(null, minified, minifier);
     },
     'should minify correctly': function(error, minified) {
-      assert.equal(minified, '');
+      assert.equal(minified.styles, '');
     },
     'should raise no errors': function(error, minified, minifier) {
       assert.equal(minifier.errors.length, 0);
@@ -234,32 +234,32 @@ vows.describe('module tests').addBatch({
       return new CleanCSS().minify(new Buffer('@import url(test/data/partials/one.css);'));
     },
     'should be processed correctly': function(minified) {
-      assert.equal('.one{color:red}', minified);
+      assert.equal('.one{color:red}', minified.styles);
     }
   },
   'options': {
     'advanced': {
       'topic': new CleanCSS({ advanced: true }).minify('a{color:red}a{color:#fff}'),
       'gets right output': function (minified) {
-        assert.equal('a{color:#fff}', minified);
+        assert.equal('a{color:#fff}', minified.styles);
       }
     },
     'aggressive merging': {
       'topic': new CleanCSS({ aggressiveMerging: true }).minify('a{display:block;color:red;display:inline-block}'),
       'gets right output': function (minified) {
-        assert.equal('a{color:red;display:inline-block}', minified);
+        assert.equal('a{color:red;display:inline-block}', minified.styles);
       }
     },
     'process import': {
       'topic': new CleanCSS({ processImport: true }).minify('@import url(/test/data/partials/one.css);'),
       'gets right output': function (minified) {
-        assert.equal('.one{color:red}', minified);
+        assert.equal('.one{color:red}', minified.styles);
       }
     },
     'rebase': {
       'topic': new CleanCSS({ rebase: true, relativeTo: path.join(process.cwd(), 'test', 'data'), root: process.cwd() }).minify('div{background:url(./dummy.png)}'),
       'gets right output': function (minified) {
-        assert.include(minified, 'url(/test/data/dummy.png)');
+        assert.include(minified.styles, 'url(/test/data/dummy.png)');
       }
     }
   }