+++ test/binary-test.js
@@ -235,9 +235,9 @@ exports.commandsSuite = vows.describe('binary commands').addBatch({
     absolute: binaryContext('-r ./test/data/129-assets --skip-rebase ./test/data/129-assets/assets/ui.css', {
       'should rebase urls correctly': function(error, stdout) {
         assert.equal(error, null);
-        assert.include(stdout, 'url(../components/bootstrap/images/glyphs.gif)');
-        assert.include(stdout, 'url(../components/jquery-ui/images/prev.gif)');
-        assert.include(stdout, 'url(../components/jquery-ui/images/next.gif)');
+        assert.include(stdout, 'url(../images/glyphs.gif)');
+        assert.include(stdout, 'url(../images/prev.gif)');
+        assert.include(stdout, 'url(../images/next.gif)');
       }
     })
   },