+++ test/protocol-imports-test.js
@@ -215,6 +215,27 @@ vows.describe('protocol imports').addBatch({
       nock.cleanAll();
     }
   },
+  'of an existing file with relative URLs and rebase turned off': {
+    topic: function() {
+      this.reqMocks = nock('http://127.0.0.1')
+        .get('/base.css')
+        .reply(200, '@import url(deeply/nested/urls.css);')
+        .get('/deeply/nested/urls.css')
+        .reply(200, 'a{background:url(../images/test.png)}');
+
+      new CleanCSS({ rebase: false }).minify('@import url(http://127.0.0.1/base.css);', this.callback);
+    },
+    'should not raise errors': function(errors, minified) {
+      assert.isNull(errors);
+    },
+    'should process @import': function(errors, minified) {
+      assert.equal(minified.styles, 'a{background:url(../images/test.png)}');
+    },
+    teardown: function() {
+      assert.equal(this.reqMocks.isDone(), true);
+      nock.cleanAll();
+    }
+  },
   'of a non-resolvable domain': {
     topic: function() {
       new CleanCSS().minify('@import url(http://notdefined.127.0.0.1/custom.css);a{color:red}', this.callback);